+++ .idea/modules.xml
@@ -5,6 +5,7 @@
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/ChatBot_MET_CS622.iml" filepath="$PROJECT_DIR$/.idea/modules/ChatBot_MET_CS622.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/app/ChatBot_MET_CS622.app.iml" filepath="$PROJECT_DIR$/.idea/modules/app/ChatBot_MET_CS622.app.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/firebase-api-library/ChatBot_MET_CS622.firebase-api-library.iml" filepath="$PROJECT_DIR$/.idea/modules/firebase-api-library/ChatBot_MET_CS622.firebase-api-library.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/modules/network-library/ChatBot_MET_CS622.network-library.iml" filepath="$PROJECT_DIR$/.idea/modules/network-library/ChatBot_MET_CS622.network-library.iml" />
     </modules>
   </component>
 </project>
