+++ app/build.gradle
@@ -79,12 +79,8 @@ dependencies {
     implementation "androidx.work:work-runtime-ktx:2.4.0"
     implementation 'com.google.guava:guava:30.1-android'
 
-//    Networking libraries
-    implementation 'com.google.code.gson:gson:2.8.6'
-    implementation 'com.squareup.retrofit2:retrofit:2.9.0'
-    implementation "com.squareup.retrofit2:converter-moshi:2.9.0"
-
     implementation project(':firebase-api-library')
+    implementation project(':network-library')
     implementation "androidx.core:core-ktx:1.3.2"
     implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.2.0"
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
