+++ network-library/src/main/java/com/example/network_library/Constants.kt
@@ -1,4 +1,4 @@
-package com.fazemeright.chatbotmetcs622.network
+package com.example.network_library
 
 object Constants {
     const val BASE_APP_NAME = "/MET_CS622_ChatBot_Backend_war_exploded"
